+++ schema/macaddr_array.go
@@ -56,6 +56,7 @@ func (dst *MacaddrArray) String() string {
 			}
 			sb.WriteString(element.String())
 		}
+		sb.WriteString("}")
 	} else {
 		return ""
 	}
